refactor(onepoint): extract deriveEnRouteHelpers into pickup-enroute.ts
Move the en-route helper derivation out of the NowHub inline block into a
small, pure exported function in pickup-enroute.ts. The test now imports the
real function instead of a local mirror copy, eliminating the false-confidence
gap flagged in the review finding. Behaviour is identical.
